separate out models from app.py, changed API URI, and created helper functions
